Move peer scripts into proper folder
