+++ lib/peer/default-peer.js
@@ -1,11 +1,11 @@
 const Peer = require('./peer.js')
 const zlib = require('zlib')
-const Messages = require('./messages.js')
-const MessageFactory = require('./message-factory.js')
+const Messages = require('../messages.js')
+const MessageFactory = require('../message-factory.js')
 const downloadPeerFile = require('./download-peer-file.js')
 const debug = require('debug')('slsk:peer:default:i')
 
-let stack = require('./stack')
+let stack = require('../stack')
 
 module.exports = class DefaultPeer extends Peer {
   constructor (socket, peer) {
