+++ lib/peer/distributed-peer.js
@@ -1,6 +1,6 @@
 const Peer = require('./peer.js')
-const Messages = require('./messages.js')
-const MessageFactory = require('./message-factory.js')
+const Messages = require('../messages.js')
+const MessageFactory = require('../message-factory.js')
 const debug = require('debug')('slsk:peer:distributed:i')
 
 module.exports = class DistributedPeer extends Peer {
