+++ lib/peer/download-peer-file.js
@@ -1,9 +1,9 @@
 const net = require('net')
 const fs = require('fs')
 const debug = require('debug')('slsk:peer:file')
-const MessageFactory = require('./message-factory.js')
+const MessageFactory = require('../message-factory.js')
 
-let stack = require('./stack')
+let stack = require('../stack')
 
 module.exports = (host, port, token, user, noPierce) => {
   debug(`downloadPeerFile ${user}`)
