+++ lib/peer/peer.js
@@ -1,5 +1,5 @@
 const EventEmitter = require('events')
-const MessageFactory = require('./message-factory.js')
+const MessageFactory = require('../message-factory.js')
 const debug = require('debug')('slsk:peer:i')
 
 module.exports = class DefaultPeer extends EventEmitter {
