+++ lib/slsk-client.js
@@ -5,10 +5,10 @@ const net = require('net')
 
 const debug = require('debug')('slsk:i')
 
-const downloadPeerFile = require('./download-peer-file.js')
 const Server = require('./server.js')
-const DefaultPeer = require('./default-peer.js')
-const DistributedPeer = require('./distributed-peer.js')
+const DefaultPeer = require('./peer/default-peer.js')
+const DistributedPeer = require('./peer/distributed-peer.js')
+const downloadPeerFile = require('./peer/download-peer-file.js')
 const Listen = require('./listen.js')
 let stack = require('./stack')
